Move AnchorTarget to tree.
