+++ ftml/src/enums.rs
@@ -22,77 +22,6 @@ use std::borrow::Cow;
 use std::convert::TryFrom;
 use strum_macros::IntoStaticStr;
 
-#[derive(
-    Serialize, Deserialize, IntoStaticStr, Debug, Copy, Clone, Hash, PartialEq, Eq,
-)]
-#[serde(rename_all = "kebab-case")]
-pub enum AnchorTarget {
-    /// Open the link in a new tab.
-    /// HTML attribute is `_blank`.
-    NewTab,
-
-    /// Open the link in the parent frame.
-    /// HTML attribute is `_parent`.
-    Parent,
-
-    /// Open the link in the top-most frame.
-    /// HTML attribute is `_top`.
-    Top,
-
-    /// Open the link in the current frame.
-    /// HTML attribute is `_self`.
-    /// This is the default setting, so the "anchor" field does not need to be included.
-    Same,
-}
-
-impl AnchorTarget {
-    #[inline]
-    pub fn name(self) -> &'static str {
-        self.into()
-    }
-
-    #[inline]
-    pub fn html_attr(self) -> &'static str {
-        match self {
-            AnchorTarget::NewTab => "_blank",
-            AnchorTarget::Parent => "_parent",
-            AnchorTarget::Top => "_top",
-            AnchorTarget::Same => "_same",
-        }
-    }
-
-    #[inline]
-    pub fn html_attr_needed(self) -> Option<&'static str> {
-        match self {
-            AnchorTarget::NewTab => Some("_blank"),
-            AnchorTarget::Parent => Some("_parent"),
-            AnchorTarget::Top => Some("_top"),
-            AnchorTarget::Same => None,
-        }
-    }
-}
-
-impl<'a> TryFrom<&'a str> for AnchorTarget {
-    type Error = ();
-
-    fn try_from(value: &'a str) -> Result<Self, Self::Error> {
-        const ANCHOR_TARGET_VALUES: [(&str, &str, AnchorTarget); 4] = [
-            ("blank", "_blank", AnchorTarget::NewTab),
-            ("parent", "_parent", AnchorTarget::Parent),
-            ("top", "_top", AnchorTarget::Top),
-            ("self", "_self", AnchorTarget::Same),
-        ];
-
-        for (value1, value2, target) in &ANCHOR_TARGET_VALUES {
-            if value.eq_ignore_ascii_case(value1) || value.eq_ignore_ascii_case(value2) {
-                return Ok(*target);
-            }
-        }
-
-        Err(())
-    }
-}
-
 #[derive(Serialize, Deserialize, Debug, Copy, Clone, Hash, PartialEq, Eq)]
 #[serde(rename_all = "kebab-case")]
 pub enum HeadingLevel {
