+++ ftml/src/parsing/rule/impls/link_single.rs
@@ -25,7 +25,8 @@
 //! Its syntax is `[https://example.com/ Label text]`.
 
 use super::prelude::*;
-use crate::enums::{AnchorTarget, LinkLabel};
+use crate::enums::LinkLabel;
+use crate::tree::AnchorTarget;
 
 pub const RULE_LINK_SINGLE: Rule = Rule {
     name: "link-single",
