+++ ftml/src/parsing/rule/impls/link_triple.rs
@@ -29,7 +29,8 @@
 //! Its syntax is `[[[page-name | Label text]`.
 
 use super::prelude::*;
-use crate::enums::{AnchorTarget, LinkLabel};
+use crate::enums::LinkLabel;
+use crate::tree::AnchorTarget;
 
 pub const RULE_LINK_TRIPLE: Rule = Rule {
     name: "link-triple",
